+++ src/main/java/woowacourse/shoppingcart/dao/CustomerDao.java
@@ -5,13 +5,13 @@ import woowacourse.shoppingcart.entity.CustomerEntity;
 public interface CustomerDao {
     int save(CustomerEntity customerEntity);
 
-    CustomerEntity findById(int id);
+    CustomerEntity findById(long id);
 
     CustomerEntity findByEmail(String email);
 
-    void update(int id, CustomerEntity customerEntity);
+    void update(long id, CustomerEntity customerEntity);
 
-    void delete(int id);
+    void delete(long id);
 
     boolean existsById(long id);
